type (isStacK -> isStack)
saw this right above where the other thing was. Not related to this issue but still its own, I would think.

This probably doesn't come up much because, first of all, it's sort of last-resort, so if it's a different kind of input everything's fine, and second, I'm not sure how often stringify is even used (only for the generator? and if so, no Scratch project actually has this stuff in it)