Rework define hat recognition
Previously we were assuming that the `define` keyword was a single word
in all languages and that it was always a prefix. We were matching
`define` hats before matching built-in blocks. In fact, we were using a
totally separate parser for define hats!

I've removed the define hat parser (the `pOutline` function), and moved
`define` matching to a later stage. The result is less clean, but we can
now support languages where the define is a suffix rather than a prefix,
and we can match built-in blocks before checking for a define hat, which
are both highly desirable. I've also added support for multiple-word
prefixes and suffixes.

As a side-effect of this change we will now handle certain edge cases
differently. For example, previously we would parse `[foo v]` inside a
define hat as a string reporter with the label "foo v"; it will now be
parsed as a dropdown. I don't think breaking backwards-compatibility is
too bad here; the syntax for define hats has never been particularly
well-defined in the first place!

This also breaks compatibility with old langauge packs (suggesting we
need a new minor release, i.e. v3.5).

Resolves #311.

Resolves #266 (since we no longer call `paintBlock` when processing the
outline).