Avoid `use xlink:href` in SVG
Fixes #348.

Currently if we export an SVG, the `use` elements become the following:

```html
<use xmlns:ns1="http://www.w3.org/1999/xlink" ns1:href="#sb3-greenFlag" ... />
```

This apparently causes a problem with Discourse and/or its virtual DOM.

This commit instead omits `xlink:href` and simply sets `href`.

MDN recommends this; they explicitly recommend against `xlink:href`. And
`<use href>` seems to be supported in all the browsers we care about,
including IE.