Don't crop puzzle piece from trailing C blocks
If a c-block was the last thing in a script, its bottom puzzle piece would get chopped off.

I believe this is because the test for `hasPuzzle` was wrong.

While this code is shared between both the `scratch2` and `scratch3` renderers, it doesn't seem to affect the Scratch 2.0 renderer. In fact, this flag is only used in exactly one place: in the Scratch 3 `ScriptView` renderer!