syntax: Make comparator predicate detection more specific
Fixes #399.

We have a special rule in the parser that detects the comparison predicates `<()>()>` and `<()<()>` and makes sure the central <> is parsed as a label, rather than the start or end of a predicate.

This rule was triggering in a bunch of situations we didn't want. This
change makes it more specific, so that it should be (almost) impossible
to trigger when we don't want.