scratch2: Handle newlines in translatePath
We were failing to render Scratch 2 rings in some cases.

This happened because we that renderer has a function `translatePath` which shifts the individual X/Y coordinates in a path string. This function was splitting on spaces, but we recently switched the path generation functions to use template strings, I suspect the extra newlines were messing things up.

Splitting the path on any whitespace, rather than just spaces, seems to fix the assertion failure.

The Scratch 3 renderer doesn't have a `translatePath` function, so it doesn't have this problem.