development: Add screenshot overlay
This commit adds the option, when developing locally, to overlay the file `screenshot.png` on top of the output from Scratchblocks. This can be used to spot differences in layout and pixel alignment compared with Scratch.

Grey pixels on the top and left of the screenshot image are automatically cut off, so it's easier to get the alignment correct.